--- a/build.fsx
+++ b/build.fsx
@@ -210,6 +210,7 @@ Target "CreateNuget" (fun _ ->
     let projects = !! "src/**/*.csproj" 
                    -- "src/**/*Tests.csproj" // Don't publish unit tests
                    -- "src/**/*Tests*.csproj"
+                   -- "src/**/*Benchmark*.csproj"
 
     let runSingleProject project =
         DotNetCli.Pack