--- a/src/Akka.Streams.Kafka.Benchmark/Akka.Streams.Kafka.Benchmark.csproj
+++ b/src/Akka.Streams.Kafka.Benchmark/Akka.Streams.Kafka.Benchmark.csproj
@@ -8,7 +8,7 @@
 
     <ItemGroup>
       <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
-        <PackageReference Include="Docker.DotNet" Version="3.125.4" />
+        <PackageReference Include="Docker.DotNet" Version="3.125.5" />
     </ItemGroup>
 
     <ItemGroup>