--- a/src/Akka.Streams.Kafka.Benchmark/Program.cs
+++ b/src/Akka.Streams.Kafka.Benchmark/Program.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using BenchmarkDotNet.Running;
+
+namespace Akka.Streams.Kafka.Benchmark
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            BenchmarkRunner.Run<PlainSinkConsumerBenchmark>();
+        }
+    }
+}