… RELEASE_NOTES.md
… src/Akka.Streams.Kafka.Tests/Integration/PlainPartitionedManualOffsetSourceIntegrationTests.cs
… src/Akka.Streams.Kafka.Tests/Integration/PlainPartitionedSourceIntegrationTests.cs
… src/Akka.Streams.Kafka.Tests/Integration/PlainSourceIntegrationTests.cs
… src/Akka.Streams.Kafka.Tests/KafkaFixture.cs
… src/Akka.Streams.Kafka.Tests/KafkaIntegrationTests.cs
… src/Akka.Streams.Kafka/Helpers/RebalanceListenerBase.cs
… src/Akka.Streams.Kafka/Stages/Consumers/Abstract/SingleSourceStageLogic.cs
… src/Akka.Streams.Kafka/Stages/Consumers/Abstract/SubSourceLogic.cs
… src/Akka.Streams.Kafka/Stages/Consumers/Actors/KafkaConsumerActor.cs
… src/common.props